+++ xlra/src/main/java/com/moorkensam/xlra/service/util/LogRecordFactory.java
@@ -1,5 +1,6 @@
 package com.moorkensam.xlra.service.util;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.moorkensam.xlra.model.configuration.LogRecord;
@@ -8,7 +9,7 @@ import com.moorkensam.xlra.model.configuration.RateLogRecord;
 
 public class LogRecordFactory {
 
-	public static LogRecord createChfLogRecord(double value) {
+	public static LogRecord createChfLogRecord(BigDecimal value) {
 		RateLogRecord record = new RateLogRecord();
 		fillInBasicProperties(record);
 		record.setRate(value);
@@ -20,7 +21,7 @@ public class LogRecordFactory {
 		record.setLogDate(new Date());
 	}
 
-	public static LogRecord createDieselLogRecord(double value) {
+	public static LogRecord createDieselLogRecord(BigDecimal value) {
 		RateLogRecord record = new RateLogRecord();
 		fillInBasicProperties(record);
 		record.setType(LogType.DIESELRATE);
